// todo_frontend/src/types/ActionType.ts
export const enum ActionType {
  ADD_TODO = 'ADD_TODO',
  ADD_TODO_REQUEST = 'ADD_TODO_REQUEST',
  ADD_TODO_SUCCESS = 'ADD_TODO_SUCCESS',
  DELETE_TODO = 'DELETE_TODO',
  DELETE_TODO_REQUEST = 'DELETE_TODO_REQUEST',
  DELETE_TODO_SUCCESS = 'DELETE_TODO_SUCCESS',
  TOGGLE_FINISHED = 'TOGGLE_FINISHED',
  TOGGLE_FINISHED_REQUEST = 'TOGGLE_FINISHED_REQUEST',
  TOGGLE_FINISHED_SUCCESS = 'TOGGLE_FINISHED_SUCCESS',
  RENEW_TODO_LIST = 'RENEW_TODO_LIST',
}